S11/docker-flask/requirements.txt
flask==3.1
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
